code freeze for stable release build.3
Nice work guys. Seems to be almost all is sorted out.

Well not all devices will get a stable:

g2 for sure will not

p3xx and p5xx possible as well not

all other seems to be on track

Code is closed now. If urgent device specific changes need to be pushed then push
and notify Jenkins to rebuild the build for the tester.

Will start now jenkins to build all builds so that they can get to our testing
crew

Change-Id: I8587afb4ad8052ad3e47acc5eb2572e6ae1d9ed4