Update patch set 2
Patch Set 2: Code-Review-1

I always clean flash when connected to wifi so didn't think about that, but Robbie's remark is correct. Now with this commit we don't let the user to run the setup wizard without a wifi connection, that's not so good too. A switch would be the best solution.

Patch-set: 2